JarFile W/ designPics & originalUML
Attached is the Entire Jar File that I have been working with in order to develop our Memory Game Project. Class Card Allows us to create Specific Jbuttons with a FrontSide and BackSide(See Method Descriptions) Class Deck is basically finished and is a collection of class card, it is programed to an interface to allow us to use any kind of  list that we need in order to make it all happen. There is a GetCollection() and GetHashSet for the deck that will allow us to compare JButtons, The Class card will come in handy when we need to compare JButtons as we cannot directly compare Jbuttons. Main class is simply a test Client I have been using to test out certain areas of the code. Main Gui is where the Main Method truly exists. Before you edit or change anything, please run the Main Gui to see where we are currently. Follows the pictures in the folder if you get lost. I have done my best to comment where I feel is neccessary. Don't hesitate to text me or message me with any questions 8019464644. Player Class is already finished and doesn't need anything updated. Player Type is fine the way it is as well. PracticeGui class is what I am using in order to build Jpanels before I add them to the main gui . Once again, follow the pictures and you should be able to follow along. Selection Screen is also a test class for Jpanels. Do your best to approach the project based off the pictures. make adjustments where you feel neccessary and we will compare as we go on . Once Again, let me know if you have any questions.